Add basic entity CoffeeBreak, move entities Campus and University to package university
