+++ social-service/src/main/java/com/eirsteir/coffeewithme/social/repository/UniversityRepository.java
@@ -1,6 +1,6 @@
 package com.eirsteir.coffeewithme.social.repository;
 
-import com.eirsteir.coffeewithme.social.domain.University;
+import com.eirsteir.coffeewithme.social.domain.university.University;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UniversityRepository extends JpaRepository<University, Long> {
