+++ social-service/src/main/java/com/eirsteir/coffeewithme/social/web/request/UpdateProfileRequest.java
@@ -21,7 +21,5 @@ public class UpdateProfileRequest {
     @NotBlank(message = "University id is required")
     private Long universityId;
 
-    @NotBlank(message = "University id is required")
-    private Long universityId;
 
 }
